feat: 리뷰 반영 - Slider component 분리
